+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @byrding/react
 
+## 0.2.0
+
+### Minor Changes
+
+- Export `CoreStore` class for multi-core isolation; add optional `{ core }` option to `defineStore` in React and Vue adapters. Rename internal `CoreStore<T>` interface to `StoreHandle<T>` to resolve naming conflict.
+
+### Patch Changes
+
+- Updated dependencies
+  - @byrding/core@0.4.0
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @byrding/core@0.3.0
+
 ## 0.1.3
 
 ### Patch Changes